Merge master into ObservingList

+++ src/scripting/StelScriptMgr.cpp
@@ -37,6 +37,7 @@
 
 #include "StelSkyDrawer.hpp"
 #include "StelSkyLayerMgr.hpp"
+#include "StelUtils.hpp"
 
 #include <QDateTime>
 #include <QDebug>
@@ -52,22 +53,43 @@
 
 #include <cmath>
 
-QScriptValue vec3fToScriptValue(QScriptEngine *engine, const Vec3f& c)
+// 3f - 3f - 3f - 3f - 3f - 3f - 3f - 3f - 3f - 3f - 3f - 3f - 3f - 3f
+
+/***
+The C++ type Vec3f is an array of three floats, used for representing
+colors according to the terms of the RGB color model. It is mapped to
+a JavaScript class with a constructor Vec3f, properties 'r', 'g' and 'b',
+and methods toString and toHex. The latter returns a string according to
+the pattern '#hhhhhh' (six hexadecimal digits). A second constructor,
+Color, is provided as a convenience. Both constructors may be called without
+argument to return  'white', a single argument which is either a color name
+or a '#hhhhhh' string, or three arguments, specifying the RGB values,
+floating point numbers between 0 and 1. Color names correspond to the names
+defined in https://www.w3.org/TR/SVG11/types.htm.
+***/
+
+QScriptValue vec3fToString(QScriptContext* context, QScriptEngine *engine)
 {
-	QScriptValue obj = engine->newObject();
-	obj.setProperty("r", QScriptValue(engine, static_cast<qsreal>(c[0])));
-	obj.setProperty("g", QScriptValue(engine, static_cast<qsreal>(c[1])));
-	obj.setProperty("b", QScriptValue(engine, static_cast<qsreal>(c[2])));
-	return obj;
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	QScriptValue rVal = that.property( "r", QScriptValue::ResolveLocal );
+	QScriptValue gVal = that.property( "g", QScriptValue::ResolveLocal );
+	QScriptValue bVal = that.property( "b", QScriptValue::ResolveLocal );
+	return "[r:" + rVal.toString() + ", " + "g:" + gVal.toString() + ", " +
+            "b:" + bVal.toString() + "]";
 }
 
-QScriptValue vec3dToScriptValue(QScriptEngine *engine, const Vec3d& c)
+QScriptValue vec3fToHex(QScriptContext* context, QScriptEngine *engine)
 {
-	QScriptValue obj = engine->newObject();
-	obj.setProperty("r", QScriptValue(engine, static_cast<qsreal>(c[0])));
-	obj.setProperty("g", QScriptValue(engine, static_cast<qsreal>(c[1])));
-	obj.setProperty("b", QScriptValue(engine, static_cast<qsreal>(c[2])));
-	return obj;
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	QScriptValue rVal = that.property( "r", QScriptValue::ResolveLocal );
+	QScriptValue gVal = that.property( "g", QScriptValue::ResolveLocal );
+	QScriptValue bVal = that.property( "b", QScriptValue::ResolveLocal );
+	return QString("#%1%2%3")
+		.arg(qMin(255, int(rVal.toNumber() * 255)), 2, 16, QChar('0'))
+		.arg(qMin(255, int(gVal.toNumber() * 255)), 2, 16, QChar('0'))
+		.arg(qMin(255, int(bVal.toNumber() * 255)), 2, 16, QChar('0'));
 }
 
 void vec3fFromScriptValue(const QScriptValue& obj, Vec3f& c)
@@ -77,28 +99,199 @@ void vec3fFromScriptValue(const QScriptValue& obj, Vec3f& c)
 	c[2] = static_cast<float>(obj.property("b").toNumber());
 }
 
-void vec3dFromScriptValue(const QScriptValue& obj, Vec3d& c)
+QScriptValue vec3fToScriptValue(QScriptEngine *engine, const Vec3f& c)
 {
-	c[0] = static_cast<double>(obj.property("r").toNumber());
-	c[1] = static_cast<double>(obj.property("g").toNumber());
-	c[2] = static_cast<double>(obj.property("b").toNumber());
+	QScriptValue obj = engine->newObject();
+	obj.setProperty("r", QScriptValue(engine, static_cast<qsreal>(c[0])));
+	obj.setProperty("g", QScriptValue(engine, static_cast<qsreal>(c[1])));
+	obj.setProperty("b", QScriptValue(engine, static_cast<qsreal>(c[2])));
+	obj.setProperty("toString", engine->newFunction( vec3fToString ));
+	obj.setProperty("toHex",    engine->newFunction( vec3fToHex ));
+	return obj;
 }
 
+// Constructor Vec3f
+// Three arguments are r, g, b. One argument is "#hhhhhh", or, perhaps, a color name.
 QScriptValue createVec3f(QScriptContext* context, QScriptEngine *engine)
 {
 	Vec3f c;
-	c[0] = static_cast<float>(context->argument(0).toNumber());
-	c[1] = static_cast<float>(context->argument(1).toNumber());
-	c[2] = static_cast<float>(context->argument(2).toNumber());
+	switch( context->argumentCount() )
+	{
+		case 0:
+			// no color: black or white? Let's say: white, against a black sky.
+			c.set( 1, 1, 1 );
+			break;
+		case 1:
+			// either '#hhhhhh' or a color name - let QColor do the work
+			if( context->argument(0).isString() )
+			{
+				QColor qcol = QColor( context->argument(0).toString() );
+				if( qcol.isValid() )
+				{
+					c.set( static_cast<float>(qcol.redF()), static_cast<float>(qcol.greenF()), static_cast<float>(qcol.blueF()) );
+					break;
+				}
+				else
+					context->throwError( QString("Color: invalid color name") );
+			}
+			else
+			{
+				// Let's hope: a Color/Vec3f object
+				return context->argument(0);
+			}
+			break;
+		case 3:
+			// r, g, b: between 0 and 1.
+			if( context->argument(0).isNumber() &&
+			    context->argument(1).isNumber() &&
+			    context->argument(2).isNumber() )
+			{
+				c[0] = static_cast<float>(context->argument(0).toNumber());
+				c[1] = static_cast<float>(context->argument(1).toNumber());
+				c[2] = static_cast<float>(context->argument(2).toNumber());
+				if( c[0] < 0 || 1 < c[0] ||
+				    c[1] < 0 || 1 < c[1] ||
+				    c[2] < 0 || 1 < c[2] )
+				{
+					context->throwError( QString("Color: RGB value out of range [0,1]") );
+				}
+			}
+			break;
+		default:
+			context->throwError( QString("Color: invalid number of arguments") );
+	}
 	return vec3fToScriptValue(engine, c);
+}
+
+QScriptValue createColor(QScriptContext* context, QScriptEngine *engine)
+{
+	return engine->globalObject().property("Vec3f").construct(context->argumentsObject());
+}
+
+// 3d - 3d - 3d - 3d - 3d - 3d - 3d - 3d - 3d - 3d - 3d - 3d - 3d - 3d
+
+/***
+The C++ type Vec3d is an array of three doubles, used for representing
+vectors (points or directions) in three-dimensional space. The coordinate
+system is implied by the context.  It is mapped to a JavaScript class with
+the constructor Vec3d, properties 'x', 'y' and 'z' (and, for historical
+reasons, also named 'r', 'g' and 'b') and a method toString. The 
+constructor may be called without an argument, returning the null vector,
+with three arguments to set the three properties, and with two arguments,
+interpreted to mean longtitude and latitude angles (or azimuth and
+altitude angles). These angles may be given as double values for degrees,
+or as string values denoting angles according to the patterns accepted
+by StelUtils::getDecAngle. 
+***/
+
+
+QScriptValue vec3dToString(QScriptContext* context, QScriptEngine *engine)
+{
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	QScriptValue xVal = that.property( "r", QScriptValue::ResolveLocal );
+	QScriptValue yVal = that.property( "g", QScriptValue::ResolveLocal );
+	QScriptValue zVal = that.property( "b", QScriptValue::ResolveLocal );
+	return "[" + xVal.toString() + ", " + yVal.toString() + ", " +
+		         zVal.toString() + "]";
+}
+
+QScriptValue getX(QScriptContext* context, QScriptEngine *engine)
+{
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	return that.property( "r", QScriptValue::ResolveLocal );
+}
+QScriptValue getY(QScriptContext* context, QScriptEngine *engine)
+{
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	return that.property( "g", QScriptValue::ResolveLocal );
+}
+QScriptValue getZ(QScriptContext* context, QScriptEngine *engine)
+{
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	return that.property( "b", QScriptValue::ResolveLocal );
+}
+
+QScriptValue setX(QScriptContext* context, QScriptEngine *engine)
+{
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	that.setProperty("r", context->argument(0).toNumber());
+	return QScriptValue();
+}
+QScriptValue setY(QScriptContext* context, QScriptEngine *engine)
+{
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	that.setProperty("g", context->argument(0).toNumber());
+	return QScriptValue();
+}
+QScriptValue setZ(QScriptContext* context, QScriptEngine *engine)
+{
+	Q_UNUSED(engine)
+	QScriptValue that = context->thisObject();
+	that.setProperty("b", context->argument(0).toNumber());
+	return QScriptValue();
+}
+
+void vec3dFromScriptValue(const QScriptValue& obj, Vec3d& c)
+{
+	c[0] = obj.property("r").toNumber();
+	c[1] = obj.property("g").toNumber();
+	c[2] = obj.property("b").toNumber();
+}
+
+QScriptValue vec3dToScriptValue(QScriptEngine *engine, const Vec3d& v)
+{
+	QScriptValue obj = engine->newObject();
+	obj.setProperty("r", QScriptValue(engine, static_cast<qsreal>(v[0])));
+	obj.setProperty("g", QScriptValue(engine, static_cast<qsreal>(v[1])));
+	obj.setProperty("b", QScriptValue(engine, static_cast<qsreal>(v[2])));
+	obj.setProperty("toString", engine->newFunction( vec3dToString ) );
+	obj.setProperty("x", engine->newFunction(getX), QScriptValue::PropertyGetter);
+	obj.setProperty("x", engine->newFunction(setX), QScriptValue::PropertySetter);
+	obj.setProperty("y", engine->newFunction(getY), QScriptValue::PropertyGetter);
+	obj.setProperty("y", engine->newFunction(setY), QScriptValue::PropertySetter);
+	obj.setProperty("z", engine->newFunction(getZ), QScriptValue::PropertyGetter);
+	obj.setProperty("z", engine->newFunction(setZ), QScriptValue::PropertySetter);	
+	return obj;
 }
 
 QScriptValue createVec3d(QScriptContext* context, QScriptEngine *engine)
 {
 	Vec3d c;
-	c[0] = static_cast<double>(context->argument(0).toNumber());
-	c[1] = static_cast<double>(context->argument(1).toNumber());
-	c[2] = static_cast<double>(context->argument(2).toNumber());
+	switch( context->argumentCount() )
+	{
+		case 0:
+			c.set( 0, 0, 0 );
+			break;
+		case 2:
+			// longitude/azimuth, latitude/altitude - maybe string dms, maybe numeric degrees
+			double lng;
+			if( context->argument(0).isString() )
+				lng = StelUtils::getDecAngle( context->argument(0).toString() );
+			else
+				lng = static_cast<double>(context->argument(0).toNumber())*M_PI_180;
+
+			double lat;
+			if ( context->argument(1).isString())
+				lat = StelUtils::getDecAngle( context->argument(1).toString() );
+			else
+				lat = static_cast<double>(context->argument(1).toNumber())*M_PI_180;
+
+			StelUtils::spheToRect( lng, lat, c );
+			break;
+		case 3:
+			c[0] = static_cast<double>(context->argument(0).toNumber());
+			c[1] = static_cast<double>(context->argument(1).toNumber());
+			c[2] = static_cast<double>(context->argument(2).toNumber());
+			break;
+		default:
+			context->throwError( QString("Vec3d: invalid number of arguments") );
+	}
 	return vec3dToScriptValue(engine, c);
 }
 
@@ -117,8 +310,23 @@ private:
 	bool isPaused;
 };
 
+void StelScriptMgr::defVecClasses(QScriptEngine *engine)
+{
+	// Allow Vec3f management in scripts
+	qScriptRegisterMetaType(engine, vec3fToScriptValue, vec3fFromScriptValue);
+	QScriptValue ctorVec3f = engine->newFunction(createVec3f);
+	engine->globalObject().setProperty("Vec3f", ctorVec3f);
+	engine->globalObject().setProperty("Color", engine->newFunction(createColor));
+
+	// Allow Vec3d management in scripts
+	qScriptRegisterMetaType(engine, vec3dToScriptValue, vec3dFromScriptValue);
+	QScriptValue ctorVec3d = engine->newFunction(createVec3d);
+	engine->globalObject().setProperty("Vec3d", ctorVec3d);
+}
+
 StelScriptMgr::StelScriptMgr(QObject *parent): QObject(parent)
 {
+	waitEventLoop = new QEventLoop();
 	engine = new QScriptEngine(this);
 	connect(&StelApp::getInstance(), SIGNAL(aboutToQuit()), this, SLOT(stopScript()), Qt::DirectConnection);
 	// Scripting images
@@ -126,14 +334,7 @@ StelScriptMgr::StelScriptMgr(QObject *parent): QObject(parent)
 	scriptImages->init();
 	StelApp::getInstance().getModuleMgr().registerModule(scriptImages);
 
-	// Allow Vec3f and Vec3d management in scripts
-	qScriptRegisterMetaType(engine, vec3fToScriptValue, vec3fFromScriptValue);
-	// Constructor
-	QScriptValue ctor = engine->newFunction(createVec3f);
-	engine->globalObject().setProperty("Vec3f", ctor);
-	qScriptRegisterMetaType(engine, vec3dToScriptValue, vec3dFromScriptValue);
-	QScriptValue ctorD = engine->newFunction(createVec3d);
-	engine->globalObject().setProperty("Vec3d", ctorD);
+	defVecClasses(engine);
 
 	// Add the core object to access methods related to core
 	mainAPI = new StelMainScriptAPI(this);
@@ -161,18 +362,14 @@ StelScriptMgr::StelScriptMgr(QObject *parent): QObject(parent)
 void StelScriptMgr::initActions()
 {
 	StelActionMgr* actionMgr = StelApp::getInstance().getStelActionManager();
-	QSignalMapper* mapper = new QSignalMapper(this);
 	for (const auto& script : getScriptList())
 	{
 		QString shortcut = getShortcut(script);
 		QString actionId = "actionScript/" + script;
-		StelAction* action = actionMgr->addAction(
-					actionId, N_("Scripts"), q_(getName(script).trimmed()), mapper, "map()", shortcut);
-		mapper->setMapping(action, script);
+		actionMgr->addAction(actionId, N_("Scripts"), q_(getName(script).trimmed()),
+				     this, [this, script] { runScript(script); }, shortcut);
 	}
-	connect(mapper, SIGNAL(mapped(QString)), this, SLOT(runScript(QString)));
 }
-
 
 StelScriptMgr::~StelScriptMgr()
 {
@@ -235,7 +432,7 @@ QString StelScriptMgr::getHeaderSingleLineCommentText(const QString& s, const QS
 		if (nameExp.exactMatch(line))
 		{
 			file.close();
-			return nameExp.capturedTexts().at(1).trimmed();
+			return nameExp.cap(1).trimmed();
 		}
 	}
 	file.close();
@@ -334,7 +531,7 @@ QString StelScriptMgr::getDescription(const QString& s) const
 		if (!inDesc && descExp.exactMatch(line))
 		{
 			inDesc = true;
-			desc = descExp.capturedTexts().at(1) + " ";
+			desc = descExp.cap(1) + " ";
 			desc.replace("\n","");
 		}
 		else if (inDesc)
@@ -344,7 +541,7 @@ QString StelScriptMgr::getDescription(const QString& s) const
 				d = "\n";
 			else if (descContExp.exactMatch(line))
 			{
-				d = descContExp.capturedTexts().at(1) + " ";
+				d = descContExp.cap(1) + " ";
 				d.replace("\n","");
 			}
 			else
@@ -397,13 +594,13 @@ bool StelScriptMgr::runScript(const QString& fileName, const QString& includePat
 	return runPreprocessedScript(preprocessedScript,fileName);
 }
 
-bool StelScriptMgr::runScriptDirect(const QString scriptId, const QString &scriptCode, const QString& includePath)
+bool StelScriptMgr::runScriptDirect(const QString scriptId, const QString &scriptCode, int &errLoc, const QString& includePath)
 {
 	QString path = includePath.isNull() ? QString(".") : includePath;
 	if( path.isEmpty() )
 		path = QStringLiteral("scripts");
 	QString processed;
-	bool ok = preprocessScript( scriptId, scriptCode, processed, path);
+	bool ok = preprocessScript( scriptId, scriptCode, processed, path, errLoc );
 	if(ok)
 		return runPreprocessedScript(processed, "<Direct script input>");
 	return false;
@@ -453,6 +650,15 @@ bool StelScriptMgr::prepareScript( QString &script, const QString &fileName, con
 
 void StelScriptMgr::stopScript()
 {
+	// Hack abusing stopScript for two different stops: it
+	// will be called again after exit from the timer loop
+	// which we kill here if it is running.
+	if( waitEventLoop->isRunning() )
+	{
+		waitEventLoop->exit( 1 );
+		return;
+	}
+	
 	if (engine->isEvaluating())
 	{
 		GETSTELMODULE(LabelMgr)->deleteAllLabels();
@@ -466,7 +672,7 @@ void StelScriptMgr::stopScript()
 		//qDebug() << msg;
 		engine->abortEvaluation();
 	}
-	scriptEnded();
+	// "Script finished..." is emitted after return from engine->evaluate().
 }
 
 void StelScriptMgr::setScriptRate(double r)
@@ -570,21 +776,27 @@ bool StelScriptMgr::strToBool(const QString& str)
 bool StelScriptMgr::preprocessFile(const QString fileName, QFile &input, QString& output, const QString& scriptDir)
 {
 	QString aText = QString::fromUtf8(input.readAll());
-	return preprocessScript( fileName, aText, output, scriptDir );
+	int errLoc;
+	return preprocessScript( fileName, aText, output, scriptDir, errLoc );
 }
 
-bool StelScriptMgr::preprocessScript(const QString fileName, const QString &input, QString &output, const QString &scriptDir)
+bool StelScriptMgr::preprocessScript(const QString fileName, const QString &input, QString &output, const QString &scriptDir, int &errLoc )
 {
 	// The one and only top-level call for expand.
 	scriptFileName = fileName;
 	num2loc = QMap<int,QPair<QString,int>>();
 	outline = 0;
 	includeSet.clear();
-	bool result = expand( fileName, input, output, scriptDir );
-	return result;
+	errLoc = -1;
+	expand( fileName, input, output, scriptDir, errLoc );
+	if( errLoc != -1 ){
+		return false;
+	} else {
+		return true;
+	}
 }
 	
-bool StelScriptMgr::expand(const QString fileName, const QString &input, QString &output, const QString &scriptDir){
+void StelScriptMgr::expand(const QString fileName, const QString &input, QString &output, const QString &scriptDir, int &errLoc){
 	QStringList lines = input.split("\n");
 	QRegExp includeRe("^include\\s*\\(\\s*\"([^\"]+)\"\\s*\\)\\s*;\\s*(//.*)?$");
 	int curline = 0;
@@ -593,7 +805,7 @@ bool StelScriptMgr::expand(const QString fileName, const QString &input, QString
 		curline++;
 		if (includeRe.exactMatch(line))
 		{
-			QString incName = includeRe.capturedTexts().at(1);
+			QString incName = includeRe.cap(1);
 			QString incPath;
 
 			// Search for the include file.  Rules are:
@@ -607,6 +819,7 @@ bool StelScriptMgr::expand(const QString fileName, const QString &input, QString
 				incPath = StelFileMgr::findFile(scriptDir + "/" + incName);
 				if (incPath.isEmpty())
 				{
+					QString fail = scriptDir + "/" + incName;
 					qWarning() << "WARNING: file not found! Let's check standard scripts directory...";
 
 					// OK, file is not exists in relative path; Let's check standard scripts directory
@@ -614,9 +827,13 @@ bool StelScriptMgr::expand(const QString fileName, const QString &input, QString
 
 					if (incPath.isEmpty())
 					{
+						fail += " or scripts/" + incName;
 						emit(scriptDebug(QString("WARNING: could not find script include file: %1").arg(QDir::toNativeSeparators(incName))));
 						qWarning() << "WARNING: could not find script include file: " << QDir::toNativeSeparators(incName);
-						return false;
+						if( errLoc == -1 ) errLoc = output.length();
+						output += line + " // <<< " + fail + " not found\n";
+						outline++;
+						continue;
 					}
 				}
 			}
@@ -634,14 +851,17 @@ bool StelScriptMgr::expand(const QString fileName, const QString &input, QString
 				{
 					qWarning() << "script include: " << QDir::toNativeSeparators(incPath);
 					QString aText = QString::fromUtf8(fic.readAll());
-					expand( incPath, aText, output, scriptDir);
+					expand( incPath, aText, output, scriptDir, errLoc );
 					fic.close();
 				}
 				else
 				{
 					emit(scriptDebug(QString("WARNING: could not open script include file for reading: %1").arg(QDir::toNativeSeparators(incPath))));
 					qWarning() << "WARNING: could not open script include file for reading: " << QDir::toNativeSeparators(incPath);
-					return false;
+				   	if( errLoc == -1 ) errLoc = output.length();
+					output += line + " // <<< " + incPath + ": cannot open\n";
+					outline++;
+					continue;
 				}
 			}
 		}
@@ -666,7 +886,7 @@ bool StelScriptMgr::expand(const QString fileName, const QString &input, QString
 			lineIdx++;
 		}
 	}
-	return true;
+	return;
 }
 
 QString StelScriptMgr::lookup( int outputPos )
@@ -692,9 +912,8 @@ QString StelScriptMgr::lookup( int outputPos )
 
 StelScriptEngineAgent::StelScriptEngineAgent(QScriptEngine *engine) 
 	: QScriptEngineAgent(engine)
-{
-	isPaused = false;
-}
+	, isPaused(false)
+	{}
 
 void StelScriptEngineAgent::positionChange(qint64 scriptId, int lineNumber, int columnNumber)
 {
